Absorb significance crate into domain module
