+++ GEMINI.md
@@ -20,8 +20,7 @@ The architecture is modularized into 26 workspace crates:
 |-------|----------------|
 | `perfgate-types` | Core domain types and stable schemas |
 | `perfgate-error` | Shared error types and categorization |
-| `perfgate-domain` | Core business logic, statistics, and paired analysis |
-| `perfgate-significance` | Statistical significance testing (Welch's t-test) |
+| `perfgate-domain` | Core business logic, statistics, significance, and paired analysis |
 | `perfgate-budget` | Budget evaluation and verdict logic |
 | `perfgate-sha256` | Minimal SHA-256 implementation for fingerprints |
 | `perfgate-host-detect` | Host fingerprinting and mismatch detection |
