--- crates/perfgate-significance/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "perfgate-significance"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-authors.workspace = true
-description = "Statistical significance testing for benchmarking"
-readme = "README.md"
-documentation = "https://docs.rs/perfgate-significance"
-keywords = ["statistics", "significance", "t-test", "benchmarking"]
-categories = ["mathematics", "development-tools"]
-
-[dependencies]
-perfgate-types.workspace = true
-statrs.workspace = true
-
-[dev-dependencies]
-proptest = "1.7"
-approx = "0.5"
-
-[lints]
-workspace = true
