+++ crates/perfgate/src/lib.rs
@@ -24,13 +24,13 @@ pub use perfgate_app as app;
 pub use perfgate_budget as budget;
 pub use perfgate_domain as domain;
 pub use perfgate_domain::paired;
+pub use perfgate_domain::significance;
 pub use perfgate_domain::stats;
 pub use perfgate_export as export;
 pub use perfgate_host_detect as host_detect;
 pub use perfgate_render as render;
 pub use perfgate_sensor as sensor;
 pub use perfgate_sha256 as sha256;
-pub use perfgate_significance as significance;
 pub use perfgate_types as types;
 pub use perfgate_types::error;
 // validation is now part of types
@@ -41,6 +41,14 @@ pub mod integrations;
 
 /// Runtime helpers for optional diagnostics and local execution support.
 pub mod runtime;
+
+/// Core I/O-free building blocks for performance-gating policy.
+pub mod core {
+    pub use perfgate_budget as budget;
+    pub use perfgate_domain::significance;
+    pub use perfgate_domain::stats;
+    pub use perfgate_sha256 as fingerprint;
+}
 
 // Common re-exports for ergonomic use
 pub mod prelude {
