+++ docs/WORKSPACE.md
@@ -10,7 +10,6 @@ This file is automatically generated by `cargo run -p xtask -- docs-sync`.
 | `perfgate-sha256` | Minimal SHA-256 implementation (no_std compatible) | 100% |
 | `perfgate-host-detect` | Host mismatch detection for CI noise reduction | 100% |
 | `perfgate-budget` | Budget evaluation logic for performance thresholds | 100% |
-| `perfgate-significance` | Statistical significance testing (Welch's t-test) | 100% |
 | `perfgate-export` | Export formats (CSV, JSONL, HTML, Prometheus) | 90% |
 | `perfgate-render` | Markdown and GitHub annotations rendering | 90% |
 | `perfgate-sensor` | Sensor report builder for cockpit integration | 90% |
@@ -40,9 +39,8 @@ graph TD
   types --> val[perfgate-types::validation]
   host[perfgate-host-detect] --> types
   types --> budget[perfgate-budget]
-  types --> sig[perfgate-significance]
+  domain --> sig[perfgate-domain::significance]
   budget --> domain[perfgate-domain]
-  sig --> domain
   domain --> adapters[perfgate-adapters]
   adapters --> app[perfgate-app]
   app --> cli[perfgate-cli]
