+++ fuzz/fuzz_targets/fuzz_budget.rs
@@ -19,11 +19,3 @@ fuzz_target!(|input: BudgetInput| {
         let _ = determine_status(regression, input.threshold, input.warn_threshold);
     }
 });
-
-
-
-
-
-
-
-
