+++ fuzz/fuzz_targets/fuzz_export_csv.rs
@@ -10,11 +10,3 @@ fuzz_target!(|data: &[u8]| {
     let s = String::from_utf8_lossy(data);
     let _ = perfgate_export::csv_escape(&s);
 });
-
-
-
-
-
-
-
-
