+++ fuzz/fuzz_targets/fuzz_host_detect.rs
@@ -37,11 +37,3 @@ fuzz_target!(|input: HostDetectInput| {
     );
     assert!(perfgate_host_detect::detect_host_mismatch(&input.current, &input.current).is_none());
 });
-
-
-
-
-
-
-
-
