+++ fuzz/fuzz_targets/fuzz_sha256.rs
@@ -17,11 +17,3 @@ fuzz_target!(|data: &[u8]| {
     assert_eq!(hash1.len(), 64);
     assert!(hash1.chars().all(|c| c.is_ascii_hexdigit()));
 });
-
-
-
-
-
-
-
-
