+++ fuzz/fuzz_targets/fuzz_stats.rs
@@ -25,11 +25,3 @@ fuzz_target!(|input: StatsInput| {
         }
     }
 });
-
-
-
-
-
-
-
-
