+++ fuzz/fuzz_targets/fuzz_validation.rs
@@ -12,11 +12,3 @@ fuzz_target!(|data: &[u8]| {
         // Should never panic, always return Ok or Err
     }
 });
-
-
-
-
-
-
-
-
