+++ fuzz/fuzz_targets/parse_config.rs
@@ -8,11 +8,3 @@ fuzz_target!(|data: &[u8]| {
         let _ = toml::from_str::<perfgate_types::ConfigFile>(s);
     }
 });
-
-
-
-
-
-
-
-
