+++ fuzz/fuzz_targets/parse_duration.rs
@@ -10,11 +10,3 @@ fuzz_target!(|data: &[u8]| {
         let _ = humantime::parse_duration(s);
     }
 });
-
-
-
-
-
-
-
-
