+++ fuzz/fuzz_targets/parse_report_receipt.rs
@@ -5,11 +5,3 @@ use libfuzzer_sys::fuzz_target;
 fuzz_target!(|data: &[u8]| {
     let _ = serde_json::from_slice::<perfgate_types::PerfgateReport>(data);
 });
-
-
-
-
-
-
-
-
