+++ mutants.toml
@@ -28,7 +28,6 @@ include_globs = [
     "crates/perfgate-domain/src/**",
     "crates/perfgate-types/src/**",
     "crates/perfgate-budget/src/**",
-    "crates/perfgate-significance/src/**",
     "crates/perfgate-app/src/**",
     "crates/perfgate-cli/src/**",
     "crates/perfgate-adapters/src/**",
