+++ tests/cucumber.rs
@@ -27,6 +27,7 @@ use perfgate_types::{
 
 // Microcrate imports for direct testing
 use perfgate_budget::{aggregate_verdict, evaluate_budget, reason_token as budget_reason_token};
+use perfgate_domain::significance::compute_significance;
 use perfgate_domain::stats::summarize_u64;
 use perfgate_error::{
     AdapterError, ConfigValidationError, IoError, PairedError, PerfgateError, StatsError,
@@ -37,7 +38,6 @@ use perfgate_host_detect::detect_host_mismatch;
 use perfgate_render::render_markdown;
 use perfgate_sensor::SensorReportBuilder;
 use perfgate_sha256::sha256_hex;
-use perfgate_significance::compute_significance;
 use perfgate_types::validation::validate_bench_name as validate_bench_name_fn;
 
 /// World struct that holds state across BDD scenario steps.
